pyproject.toml
[project]
name = "scene-cleanup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]
